requirements.txt
pythonocc-core>=7.7.2
PySide6>=6.4.0
pytest>=7.0.0
